requirements.txt
praw~=7.8.1
python-dotenv~=1.0.1
numpy~=2.2.3
pandas~=2.2.3
seaborn~=0.13.2
matplotlib~=3.10.0
scikit-learn~=1.6.1
textblob~=0.19.0
afinn~=0.1
vaderSentiment~=3.3.2